Tests/LibWeb/Text/expected/HTML/tabIndex-attribute.txt

p.tabIndex initial value: -1
h1.tabIndex initial value: -1
a.tabIndex initial value: 0
area.tabIndex initial value: 0
button.tabIndex initial value: 0
frame.tabIndex initial value: 0
iframe.tabIndex initial value: 0
input.tabIndex initial value: 0
object.tabIndex initial value: 0
select.tabIndex initial value: 0
textarea.tabIndex initial value: 0
svg.a.tabIndex initial value: 0
